+++ pages/checkin/@id/+Page.client.ts
@@ -4,5 +4,6 @@ import { useData } from "vike-react/useData";
 
 export function Page() {
     const { checkin } = useData();
+    console.log("Checkin data:", checkin);
     return h(Checkins, { checkin });
 }
