+++ pages/checkin/@id/+data.ts
@@ -1,8 +1,15 @@
 import { fetchAPIData } from "~/components/general";
+import { rockdApiURL } from "@macrostrat-web/settings";
 
 export async function data(pageContext) {
-    const id = pageContext.routeParams.id;
-    const checkin = await fetchAPIData(`/protected/checkins?checkin_id=${id}`)
-
-    return { checkin: checkin.success.data[0] };
+    const id = parseInt(pageContext.urlParsed.pathname.split("/")[2]);
+    const data = await fetch(
+            `${rockdApiURL}/protected/checkins?checkin_id=${id}`
+        ).then((response) => {
+            if (!response.ok) {
+                throw new Error(`HTTP error! status: ${response.status}`);
+            }
+            return response.json();
+        });
+    return { checkin: data.success.data[0] };
 }
