+++ pages/checkin/index.ts
@@ -11,6 +11,18 @@ import 'mapbox-gl/dist/mapbox-gl.css';
 import { MapPosition } from "@macrostrat/mapbox-utils";
 import { PanelCard } from "@macrostrat/map-interface";
 import { LithologyList } from "@macrostrat/data-components";
+import { ClientOnly } from "vike-react/ClientOnly";
+
+function Map(props) {
+  return h(
+    ClientOnly,
+    {
+      load: () => import("./map.client").then((d) => d.MapContainer),
+      fallback: h("div.loading", "Loading map..."),
+    },
+    (component) => h(component, props)
+  );
+}
 
 export function Checkins({checkin}) {
     const [overlayOpen, setOverlayOpen] = useState(false);
@@ -208,45 +220,4 @@ function observationFooter(observation) {
             h('p', {className: "notes"}, rocks.notes),
         ]),
     ]);
-}
-
-function Map({center, showMap, setShowMap}) {
-    const [style, setStyle] = useState("mapbox://styles/jczaplewski/cje04mr9l3mo82spihpralr4i");
-    const [styleText, setStyleText] = useState("Show Satelite");
-
-    const newMapPosition: MapPosition = {
-        camera: {
-          lat: center.lat,  // Latitude
-          lng: center.lng, // Longitude
-          altitude: 300000, // Altitude (height from the Earth's surface)
-        },
-      };
-
-    const sateliteStyle = 'mapbox://styles/mapbox/satellite-v9';
-    const whiteStyle = "mapbox://styles/jczaplewski/cje04mr9l3mo82spihpralr4i";
-    const whiteText = "Show White";
-    const sateliteText = "Show Satelite";
-
-    let map = h("div.map", [
-        h(MapAreaContainer, { style: {height: "93vh", top: "7vh"} },
-            [
-              h(MapView, { style: style, mapboxToken: SETTINGS.mapboxAccessToken, mapPosition: newMapPosition }, [
-                h(MapMarker, {
-                    position: center,
-                   }),
-              ]),
-            ]
-          ),
-        h('div', {className: 'banner'}, [
-            h(Icon, {className: "banner-arrow", icon: "arrow-left", iconSize: "3vh", style: {color: 'white'}, onClick: () => {
-                setShowMap(!showMap);
-              }}),
-            h(PanelCard, {className: "banner-button", onClick: () => {
-                setStyle(style == whiteStyle ? sateliteStyle : whiteStyle);
-                setStyleText(styleText == whiteText ? sateliteText : whiteText);
-            }}, styleText),
-        ]),
-    ])
-
-    return map;
 }
